requirements.txt
tensorflow-gpu==1.12.0
scipy
numpy 
nibabel
matplotlib
seaborn
time
shutil